Optimize Datagenerator & MMap
- Migrate mmap buffer stream to long
- Modify default stream type of datagen to mmap

+++ src/utils/MemoryMappedInputStream.java
@@ -9,9 +9,9 @@ import java.nio.channels.FileChannel;
 public class MemoryMappedInputStream implements BaseInputStream {
     private FileChannel fileChannel;
     private MappedByteBuffer mappedByteBuffer;
-    private int bufferSize;
+    private long bufferSize;
 
-    public MemoryMappedInputStream(int bufferSize) {
+    public MemoryMappedInputStream(long bufferSize) {
         this.bufferSize = bufferSize;
     }
